второе задание часть

+++ MathSets/classes/Base.cs
@@ -6,7 +6,7 @@ namespace MathSets
     public class Base
     {
         public static Frame MainFrame;
-        public static double StrokeThickness = 2; // Толщина границ объектов.
+        public static int StrokeThickness = 2; // Толщина границ объектов.
         public static SolidColorBrush ColorDraggableElement = Brushes.Gray; // Цвет перемещаемых элементов.
     }
 }
